Add extended block statistics to the FreeBSD devstat collector
This adds additional staistics to the FreeBSD devstat collector
so that block device monitoring is closer to having the same
data as is enjoyed on Linux. Since the data is coming from a
different collector, it's not in the correct namespace to be
a drop in for existing dashboards, but at least we can now track
these metrics. The inspriation came from the gstat_exporter
documented at:

https://blog.tyk.nu/blog/all-in-a-days-work-prometheus-gstat_exporter-and-grafana-dashboard/

However, the version of gstat used on this exporter is relatively
new and did not work on my taget platform of FreeBSD 11.4. So I
instead updated the collector to pull the relevant metrics from
devstat based on the gstat.c code in FreeBSD 11.4 and verified
that the same metrics are available in FreeBSD 13.1 as well.

New metrics exported:
 * Percentage the device was busy
 * Device command queue length
 * Transactions per Second (IOPS) per read/write/free/other type
 * Megabytes per Second throughput per read/write type
 * Kilobytes per Transaction per read/write/free type
 * Milliseconds per Transaction per read/write/other type